services/query-api/requirements.txt
# Query Processing API Dependencies

# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Authentication
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# Supabase
supabase==2.0.0

# OpenAI
openai==1.6.0

# HTTP client
httpx==0.25.2

# Utilities
python-dateutil==2.8.2
pydantic==2.5.0
tenacity==8.2.3
